+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
 members = [
-    "atl-checker",
-    "cli",
+    "cgaal-engine",
+    "cgaal-cli",
 ]
