+++ cgaal-cli/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "atl-checker-cli"
+name = "cgaal"
 version = "0.1.0"
 authors = [
     "d702e20 <d702e20@cs.aau.dk>",
@@ -26,4 +26,4 @@ tracing-subscriber = "0.2.17"
 serde_json = "1.0.83"
 regex = { version = "1", features = ["unicode-case"] }
 humantime = "2.1.0"
-atl-checker = { path = "../atl-checker", version = "0.1.0" }
+cgaal-engine = { path = "../cgaal-engine", version = "0.1.0" }
