+++ cgaal-engine/src/algorithms/certain_zero/common.rs
