+++ cgaal-engine/src/algorithms/certain_zero/search_strategy/bfs.rs
