+++ cgaal-engine/src/algorithms/certain_zero/search_strategy/instability_heuristic_search.rs
