+++ cgaal-engine/src/algorithms/certain_zero/search_strategy/linear_optimize.rs
