+++ cgaal-engine/src/algorithms/certain_zero/search_strategy/linear_programming_search.rs
