+++ cgaal-engine/src/algorithms/game_strategy/error.rs
