+++ cgaal-engine/src/algorithms/game_strategy/format.rs
