+++ cgaal-engine/src/algorithms/game_strategy/partial.rs
