+++ cgaal-engine/src/algorithms/global/mod.rs
