+++ cgaal-engine/src/algorithms/global/singlethread.rs
