+++ cgaal-engine/src/algorithms/solve_set.rs
