+++ cgaal-engine/src/analyse.rs
