+++ cgaal-engine/src/emit_count.rs
