+++ cgaal-engine/src/game_structure/eager.rs
