+++ cgaal-engine/src/game_structure/lcgs/ir/symbol_checker.rs
