+++ cgaal-engine/src/printer.rs
