+++ cgaal-engine/src/simple_edg.rs
